LayoutTests/imported/w3c/web-platform-tests/xhr/overridemimetype-edge-cases.window-expected.txt


PASS overrideMimeType() is not reset by open(), basic
PASS overrideMimeType() is not reset by open()
PASS If charset is not overridden by overrideMimeType() the original continues to be used
PASS Charset can be overridden by overrideMimeType() with a bogus charset
